refactor(webview): keep main view helpers out of schemas
