+++ packages/extension/src/webview/managers/executionHandlers.ts
@@ -2,7 +2,7 @@ import * as vscode from 'vscode';
 
 import { prepareMainViewExecutionRequest } from '@controllers/mainView/MainViewExecutionController';
 import * as logger from '@logger/logUtils';
-import type { MainViewExecuteMessage } from '@shared/schemas/mainView';
+import type { MainViewExecuteMessage } from '@shared/mainView';
 
 const CHANNEL = 'ExecutionHandlers';
 logger.initialize(CHANNEL);
